move to sentry 8-rc1 for better koa integration
