+++ package.json
@@ -25,9 +25,9 @@
     "test": "node --test"
   },
   "dependencies": {
-    "@sentry/node": "^7.113.0",
-    "@sentry/profiling-node": "^7.113.0",
-    "@sentry/utils": "^7.113.0",
+    "@sentry/node": "^8.0.0-rc.1",
+    "@sentry/profiling-node": "^8.0.0-rc.1",
+    "@sentry/utils": "^8.0.0-rc.1",
     "@types/koa-cash": "^4.1.3",
     "dotenv": "16.4.5",
     "get-urls": "12.1.0",
